🚚 chore: 유저 관련 api를 별도의 user 폴더로 변경
